[SDTTTTT] Follow the Typescript specification.
