+++ src/aspect/after.ts
@@ -1,4 +1,4 @@
-import * as vscode from 'vscode';
+import * as vscode from "vscode";
 
 /**
  * After Aspect
@@ -7,9 +7,9 @@ import * as vscode from 'vscode';
  * @author sdttttt
  */
 export function after(...args: Array<() => void>) {
-    return function (target: any , propertyKey: string, descriptor: PropertyDescriptor) {
-        for (const callback of args) {
-            callback();
-        }
-    };
+	return function (target: any , propertyKey: string, descriptor: PropertyDescriptor) {
+		for (const callback of args) {
+			callback();
+		}
+	};
 }
