+++ src/constants.ts
@@ -1,10 +1,12 @@
+"use strict";
+
 import * as vscode from "vscode";
 
 export const yinglili: string = "heiheihei.gif";
 
 export const huihui: string = "loading.gif";
 
-export const WeekBangumiCSS = "css/week_bangumi.css";
+export const WeekBangumiCSS = "css/weekBangumi.css";
 export const BangumiCSS = "css/bangumi.css";
 export const LoadingCSS = "css/loading.css";
 
@@ -17,7 +19,7 @@ let _context: vscode.ExtensionContext;
  * @author sdttttt
  */
 export function setContext(ctx: vscode.ExtensionContext) {
-    _context = ctx;
+	_context = ctx;
 }
 
 /**
@@ -27,5 +29,5 @@ export function setContext(ctx: vscode.ExtensionContext) {
  * @author sdttttt
  */
 export function getContext(): vscode.ExtensionContext {
-    return _context;
+	return _context;
 }
