+++ src/extension.ts
@@ -1,6 +1,8 @@
+"use strict";
+
 import * as vscode from "vscode";
 import BangumiView from "./views/bangumi";
-import WeekBangumiView from "./views/week_bangumi";
+import WeekBangumiView from "./views/weekBangumi";
 import { setContext } from "./constants";
 import { getConfig } from "./configuration";
 import MainIndexList from "./match/";
@@ -14,9 +16,9 @@ let isInit: boolean = false;
  * @author sdttttt
  */
 function initializer(context: vscode.ExtensionContext) {
-  if (isInit) { return; }
-  setContext(context);
-  isInit = !isInit;
+	if (isInit) { return; }
+	setContext(context);
+	isInit = !isInit;
 }
 
 /**
@@ -27,44 +29,44 @@ function initializer(context: vscode.ExtensionContext) {
  */
 export function activate(context: vscode.ExtensionContext) {
 
-  initializer(context);
+	initializer(context);
 
-  const useReminder: any = getConfig("bangumiOpen.EnableReminder");
+	const useReminder: any = getConfig("bangumiOpen.EnableReminder");
 
-  if (<boolean>useReminder) {
-    WeekBangumiView.enableBangumiUpdateReminder();
-  }
+	if (<boolean>useReminder) {
+		WeekBangumiView.enableBangumiUpdateReminder();
+	}
 
-  // The command has been defined in the package.json file
-  // Now provide the implementation of the command with registerCommand
-  // The commandId parameter must match the command field in package.json
+	// The command has been defined in the package.json file
+	// Now provide the implementation of the command with registerCommand
+	// The commandId parameter must match the command field in package.json
 
-  context.subscriptions.push(
-    vscode.commands.registerCommand(
-      "openBangumi",
-      () => BangumiView.openBangumi()
-    ),
-    vscode.commands.registerCommand(
-      "nextPage",
-      () => BangumiView.nextPage()
-    ),
-    vscode.commands.registerCommand(
-      "backPage",
-      () => BangumiView.backPage()
-    ),
-    vscode.commands.registerCommand(
-      "jumpPage",
-      () => BangumiView.jumpPage()
-    ),
-    vscode.commands.registerCommand(
-      "weekBangumi",
-      () => WeekBangumiView.openWeekBangumi()
-    ),
-    vscode.commands.registerCommand(
-      "index",
-      () => MainIndexList.openIndexList()
-    ),
-  );
+	context.subscriptions.push(
+		vscode.commands.registerCommand(
+			"openBangumi",
+			() => BangumiView.openBangumi()
+		),
+		vscode.commands.registerCommand(
+			"nextPage",
+			() => BangumiView.nextPage()
+		),
+		vscode.commands.registerCommand(
+			"backPage",
+			() => BangumiView.backPage()
+		),
+		vscode.commands.registerCommand(
+			"jumpPage",
+			() => BangumiView.jumpPage()
+		),
+		vscode.commands.registerCommand(
+			"weekBangumi",
+			() => WeekBangumiView.openWeekBangumi()
+		),
+		vscode.commands.registerCommand(
+			"index",
+			() => MainIndexList.openIndexList()
+		),
+	);
 }
 
 /**
@@ -74,5 +76,5 @@ export function activate(context: vscode.ExtensionContext) {
  * @author sdttttt
  */
 export function deactivate() {
-  WeekBangumiView.destroyReminder();
+	WeekBangumiView.destroyReminder();
 }
