+++ src/match/index.ts
@@ -1,20 +1,22 @@
-import AbstractIndexList from './index_list';
-import YearIndexList from './year_index_list';
+"use strict";
+
+import AbstractIndexList from "./indexList";
+import YearIndexList from "./yearIndexList";
 
 export default new class MainIndexList extends AbstractIndexList {
     
     protected readonly openIndexListAfter?: () => void;
 
     protected readonly list: Array<string> = [
-        "类型", "地区", "状态", "版权", "付费", "季度", "时间", "风格"
+    	"类型", "地区", "状态", "版权", "付费", "季度", "时间", "风格"
     ];
 
     protected conditionHandler(index: string): void {
-        switch (index) {
-            case "时间":
-                YearIndexList.openIndexList();
-                break;
-        }
+    	switch (index) {
+    	case "时间":
+    		YearIndexList.openIndexList();
+    		break;
+    	}
     }
 };
