+++ src/match/indexList.ts
@@ -1,4 +1,6 @@
-import * as vscode from 'vscode';
+"use strict";
+
+import * as vscode from "vscode";
 
 /**
  * Abstract index list
@@ -43,7 +45,7 @@ export default abstract class AbstractIndexList {
      * @author sdttttt
      */
     protected openBangumiHook(): void {
-        vscode.commands.executeCommand("openBangumi");
+    	vscode.commands.executeCommand("openBangumi");
     }
 
     /**
@@ -52,16 +54,16 @@ export default abstract class AbstractIndexList {
      * @author sdttttt
      */
     openIndexList(): void {
-        vscode.window.showQuickPick(this.list).then(
-            (index: string | undefined) => {
-                if (index) {
-                    this.conditionHandler(index);
+    	vscode.window.showQuickPick(this.list).then(
+    		(index: string | undefined) => {
+    			if (index) {
+    				this.conditionHandler(index);
 
-                    if (this.openIndexListAfter) 
-                        { this.openIndexListAfter(); }
-                }
-                return;
-            }
-        );
+    				if (this.openIndexListAfter) 
+    				{ this.openIndexListAfter(); }
+    			}
+    			return;
+    		}
+    	);
     }
 }
