+++ src/test/runTest.ts
@@ -1,21 +1,21 @@
-import * as path from 'path';
+import * as path from "path";
 
-import { runTests } from 'vscode-test';
+import { runTests } from "vscode-test";
 
 async function main() {
 	try {
 		// The folder containing the Extension Manifest package.json
 		// Passed to `--extensionDevelopmentPath`
-		const extensionDevelopmentPath = path.resolve(__dirname, '../../');
+		const extensionDevelopmentPath = path.resolve(__dirname, "../../");
 
 		// The path to test runner
 		// Passed to --extensionTestsPath
-		const extensionTestsPath = path.resolve(__dirname, './suite/index');
+		const extensionTestsPath = path.resolve(__dirname, "./suite/index");
 
 		// Download VS Code, unzip it and run the integration test
 		await runTests({ extensionDevelopmentPath, extensionTestsPath });
 	} catch (err) {
-		console.error('Failed to run tests');
+		console.error("Failed to run tests");
 		process.exit(1);
 	}
 }
