+++ src/test/suite/extension.test.ts
@@ -1,14 +1,16 @@
-import * as assert from 'assert';
+"use strict";
+
+import * as assert from "assert";
 
 // You can import and use all API from the 'vscode' module
 // as well as import your extension to test it
-import * as vscode from 'vscode';
+import * as vscode from "vscode";
 // import * as myExtension from '../extension';
 
-suite('Extension Test Suite', () => {
-	vscode.window.showInformationMessage('Start all tests.');
+suite("Extension Test Suite", () => {
+	vscode.window.showInformationMessage("Start all tests.");
 
-	test('Sample test', () => {
+	test("Sample test", () => {
 		assert.equal(-1, [1, 2, 3].indexOf(5));
 		assert.equal(-1, [1, 2, 3].indexOf(0));
 	});
