+++ src/test/suite/index.ts
@@ -1,20 +1,20 @@
-import * as path from 'path';
-import * as Mocha from 'mocha';
-import * as glob from 'glob';
+import * as path from "path";
+import * as Mocha from "mocha";
+import * as glob from "glob";
 
 export function run(): Promise<void> {
 	// Create the mocha test
 	const mocha = new Mocha({
-		ui: 'tdd',
+		ui: "tdd",
 	});
 
 	mocha.useColors(true);
 	mocha.timeout(5000);
 
-	const testsRoot = path.resolve(__dirname, '..');
+	const testsRoot = path.resolve(__dirname, "..");
 
 	return new Promise((c, e) => {
-		glob('**/**.test.js', { cwd: testsRoot }, (err, files) => {
+		glob("**/**.test.js", { cwd: testsRoot }, (err, files) => {
 			if (err) {
 				return e(err);
 			}
