+++ src/test/suite/utlis.test.ts
@@ -1,9 +1,11 @@
-import { toNumber, isEmptyArray, isEmptyObject } from '../../utils/type';
+"use strict";
+
+import { toNumber, isEmptyArray, isEmptyObject } from "../../utils/type";
 import { isToday, 
-    currentTimestamp, 
-    toWeekDay, 
-    toMinuteFromSecode } from '../../utils/strings';
-import { expect } from 'chai';
+	currentTimestamp, 
+	toWeekDay, 
+	toMinuteFromSecode } from "../../utils/strings";
+import { expect } from "chai";
 
 /**
  * Utils TEST
@@ -12,70 +14,70 @@ import { expect } from 'chai';
  */
 suite("Utils TEST", () => {
 
-    test("toNumber", () => {
-        const str: string = "1";
-        const num: number = toNumber(str);
+	test("toNumber", () => {
+		const str: string = "1";
+		const num: number = toNumber(str);
 
-        const undef: undefined = undefined;
+		const undef: undefined = undefined;
 
-        expect(num).to.be.equal(1);
+		expect(num).to.be.equal(1);
 
-        // tslint:disable-next-line: no-unused-expression
-        expect(toNumber(undef)).to.be.equal(0);
-    });
+		// tslint:disable-next-line: no-unused-expression
+		expect(toNumber(undef)).to.be.equal(0);
+	});
 
-    test("is empty Array TEST", () => {
-        const emptyArr: Array<any> = [];
-        const arr: Array<number> = [1, 2, 3];
-        const arr2: Array<string> = ["a", "b", "c"];
-        // tslint:disable-next-line: no-unused-expression
-        expect(isEmptyArray(emptyArr)).to.be.true;
-        // tslint:disable-next-line: no-unused-expression
-        expect(isEmptyArray(arr)).to.be.false;
-        // tslint:disable-next-line: no-unused-expression
-        expect(isEmptyArray(arr2)).to.be.false;
-    });
+	test("is empty Array TEST", () => {
+		const emptyArr: Array<any> = [];
+		const arr: Array<number> = [1, 2, 3];
+		const arr2: Array<string> = ["a", "b", "c"];
+		// tslint:disable-next-line: no-unused-expression
+		expect(isEmptyArray(emptyArr)).to.be.true;
+		// tslint:disable-next-line: no-unused-expression
+		expect(isEmptyArray(arr)).to.be.false;
+		// tslint:disable-next-line: no-unused-expression
+		expect(isEmptyArray(arr2)).to.be.false;
+	});
 
-    test("is empty Object TEST", () => {
-        const emptyObj: object = {};
-        const obj: object = { a: 1 };
+	test("is empty Object TEST", () => {
+		const emptyObj: object = {};
+		const obj: object = { a: 1 };
 
-        // tslint:disable-next-line: no-unused-expression
-        expect(isEmptyObject(emptyObj)).to.be.true;
-        // tslint:disable-next-line: no-unused-expression
-        expect(isEmptyObject(obj)).to.be.false;
-    });
+		// tslint:disable-next-line: no-unused-expression
+		expect(isEmptyObject(emptyObj)).to.be.true;
+		// tslint:disable-next-line: no-unused-expression
+		expect(isEmptyObject(obj)).to.be.false;
+	});
 
-    test("today TEST", () => {
-        const currentTime: Date = new Date();
-        const today: string = (currentTime.getMonth() + 1) + "-" + currentTime.getDate();
+	test("today TEST", () => {
+		const currentTime: Date = new Date();
+		const today: string = (currentTime.getMonth() + 1) + "-" + currentTime.getDate();
 
-        // tslint:disable-next-line: no-unused-expression
-        expect(isToday(today)).to.be.true;
-    });
+		// tslint:disable-next-line: no-unused-expression
+		expect(isToday(today)).to.be.true;
+	});
 
-    test("timestamp TEST", () => {
-        expect(currentTimestamp()).to.be.a("number");
-    });
+	test("timestamp TEST", () => {
+		expect(currentTimestamp()).to.be.a("number");
+	});
 
-    test("Number to Week String TEST", () => {
-        expect(toWeekDay(1)).to.be.equal("星期一");
-        expect(toWeekDay(5)).to.be.equal("星期五");
+	test("Number to Week String TEST", () => {
+		expect(toWeekDay(1)).to.be.equal("星期一");
+		expect(toWeekDay(5)).to.be.equal("星期五");
 
-        expect(toWeekDay(0)).to.be.equal("");
-        expect(toWeekDay(8)).to.be.equal("");
-    });
+		expect(toWeekDay(0)).to.be.equal("");
+		expect(toWeekDay(8)).to.be.equal("");
+	});
 
-    test("To Minute From Secode TEST", () => {
-        expect(toMinuteFromSecode(60)).to.be.equal(1);
-        expect(toMinuteFromSecode(30)).to.be.equal(1);
-        expect(toMinuteFromSecode(5)).to.be.equal(1);
+	test("To Minute From Secode TEST", () => {
+		expect(toMinuteFromSecode(60)).to.be.equal(1);
+		expect(toMinuteFromSecode(30)).to.be.equal(1);
+		expect(toMinuteFromSecode(5)).to.be.equal(1);
 
-        expect(toMinuteFromSecode(120)).to.be.equal(2);
-        expect(toMinuteFromSecode(119)).to.be.equal(2);
+		expect(toMinuteFromSecode(120)).to.be.equal(2);
+		expect(toMinuteFromSecode(119)).to.be.equal(2);
 
-        expect(toMinuteFromSecode(121)).to.be.equal(3);
-    });
+		expect(toMinuteFromSecode(121)).to.be.equal(3);
+	});
 
     
 });
