+++ src/utils/file.ts
@@ -1,3 +1,5 @@
+"use strict";
+
 import * as vscode from "vscode";
 import * as path from "path";
 import { getContext } from "../constants";
@@ -10,8 +12,8 @@ import { getContext } from "../constants";
  * @author sdttttt 
  */
 export function getResourceFile(filename: string): vscode.Uri {
-    const onDiskPath = vscode.Uri.file(
-        path.join(getContext().extensionPath, 'resources', filename)
-    );
-    return onDiskPath.with({ scheme: 'vscode-resource' });
+	const onDiskPath = vscode.Uri.file(
+		path.join(getContext().extensionPath, "resources", filename)
+	);
+	return onDiskPath.with({ scheme: "vscode-resource" });
 }
