+++ src/utils/strings.ts
@@ -1,5 +1,7 @@
-import { WeekBangumiData, BangumisData } from '../request/structure';
-import { toNumber } from './type';
+"use strict";
+
+import { WeekBangumiData, BangumisData } from "../request/structure";
+import { toNumber } from "./type";
 
 /**
  * To week day
@@ -9,24 +11,24 @@ import { toNumber } from './type';
  * @author sdttttt
  */
 export function toWeekDay(day: number): string {
-    switch (day) {
-        case 1:
-            return "星期一";
-        case 2:
-            return "星期二";
-        case 3:
-            return "星期三";
-        case 4:
-            return "星期四";
-        case 5:
-            return "星期五";
-        case 6:
-            return "星期六";
-        case 7:
-            return "星期日";
-        default:
-            return "";
-    }
+	switch (day) {
+	case 1:
+		return "星期一";
+	case 2:
+		return "星期二";
+	case 3:
+		return "星期三";
+	case 4:
+		return "星期四";
+	case 5:
+		return "星期五";
+	case 6:
+		return "星期六";
+	case 7:
+		return "星期日";
+	default:
+		return "";
+	}
 }
 
 
@@ -38,9 +40,9 @@ export function toWeekDay(day: number): string {
  * @author sdttttt
  */
 export function isToday(day: string): boolean {
-    const currentTime: Date = new Date();
-    const today = (currentTime.getMonth() + 1) + "-" + currentTime.getDate();
-    return today === day;
+	const currentTime: Date = new Date();
+	const today = (currentTime.getMonth() + 1) + "-" + currentTime.getDate();
+	return today === day;
 }
 
 /**
@@ -51,7 +53,7 @@ export function isToday(day: string): boolean {
  * @author sdttttt
  */
 export function currentTimestamp(): number {
-    return Number(new Date());
+	return Number(new Date());
 }
 
 /**
@@ -63,11 +65,11 @@ export function currentTimestamp(): number {
  * @author sdttttt
  */
 export function getTodayIndexInWeekBangumi(bangumis: Array<WeekBangumiData>): number | undefined {
-    for (let i = 0; i < bangumis.length; i++) {
-        if (isToday(bangumis[i].date)) {
-          return i;
-        }
-    }
+	for (let i = 0; i < bangumis.length; i++) {
+		if (isToday(bangumis[i].date)) {
+			return i;
+		}
+	}
 }
 
 /**
@@ -78,7 +80,7 @@ export function getTodayIndexInWeekBangumi(bangumis: Array<WeekBangumiData>): nu
  * @author sdttttt 
  */
 export function toMinuteFromSecode(time: number): number{
-   return Math.ceil(time / 60);
+	return Math.ceil(time / 60);
 }
 
 /**
@@ -90,5 +92,5 @@ export function toMinuteFromSecode(time: number): number{
  * @author sdttttt 
  */
 export function numberOfStringPlus(str: string, num: number): string {
-    return (toNumber(str) + num).toString();
+	return (toNumber(str) + num).toString();
 }
