+++ src/utils/type.ts
@@ -1,3 +1,5 @@
+"use strict";
+
 /**
  * Detroit: Become Number
  *
@@ -9,18 +11,18 @@
  */
 export function toNumber(text: string | undefined): number {
 
-    if (text === undefined) {
-        return 0;
-    }
+	if (text === undefined) {
+		return 0;
+	}
 
-    if (typeof text === "string") {
-        const number = Number(text);
-        if (!isNaN(number)) {
-            return number;
-        }
-    }
+	if (typeof text === "string") {
+		const number = Number(text);
+		if (!isNaN(number)) {
+			return number;
+		}
+	}
 
-    return 0;
+	return 0;
 }
 
 /**
@@ -32,7 +34,7 @@ export function toNumber(text: string | undefined): number {
  * @author sdttttt
  */
 export function isEmptyArray(array: Array<any>): boolean {
-    return array.length === 0;
+	return array.length === 0;
 }
 
 /**
@@ -44,5 +46,5 @@ export function isEmptyArray(array: Array<any>): boolean {
  * @author sdttttt
  */
 export function isEmptyObject(obj: any): boolean {
-    return JSON.stringify(obj) === "{}";
+	return JSON.stringify(obj) === "{}";
 }
